[Feature] ORG 정보 수정 0.1.2
