+++ src/apis/PointAPI.js
@@ -1,5 +1,5 @@
 import Swal from 'sweetalert2';
-import { GET_EXCHANGE, GET_USER_EXCHANGES, GET_EXCHANGE_DETAIL_U, PUT_EXCHANGES, POST_EXCHANGES, GET_POINTS_HISTORY, GET_ADMIN_EXCHANGES, PUT_PROVIDE_INFO, GET_PRIVACY_STATUS, GET_ORG_INFORMATION } from '../modules/PointModule';
+import { GET_EXCHANGE, GET_USER_EXCHANGES, GET_EXCHANGE_DETAIL_U, PUT_EXCHANGES, POST_EXCHANGES, GET_POINTS_HISTORY, GET_ADMIN_EXCHANGES, PUT_PROVIDE_INFO, GET_PRIVACY_STATUS } from '../modules/PointModule';
 import axios from "axios";
 
 export const pointExchangeAPI = ({formdata}) => {
@@ -266,43 +266,6 @@ export function getPrivacyStatusAPI(memberCode) {
             dispatch({type: GET_PRIVACY_STATUS, payload: result.data});
         } catch (error){
             console.error('[ReceiptAPI] getPrivacyStatusAPI 에러 발생 : ', error);
-        }
-    }
-}
-
-export function VerifyPwdAPI({orgCode, orgPwd}){
-    console.log('[OrgAPI] verifyPwdAPI 시작');
-    const requestURL = `http://localhost:8001/orgInfo/${orgCode}`;
-    return async (dispatch, getState) => {
-        try {
-            console.log("비밀번호? ", orgPwd);
-            const result = await axios.get(requestURL, {params : {orgPwd}});
-            console.log('[OrgAPI] verifyPwdAPI RESULT : ', result.data);
-            dispatch({type: GET_ORG_INFORMATION, payload: result.data});
-            localStorage.setItem("orgData", JSON.stringify(result.data[0]));
-            window.location.href='/editOrg/modifyOrg';
-        } catch (error){
-            console.error('[OrgAPI] verifyPwdAPI 에러 발생 : ', error);
-            if(error.response?.status == 400){
-                Swal.fire({
-                    icon: "error",
-                    iconColor: "#DB524E",
-                    title: "비밀번호를 다시 확인해 주세요.",
-                    showCancelButton: false,
-                    confirmButtonColor: '#1D7151',
-                    confirmButtonText: '확인'
-                })
-            }else{
-                Swal.fire({
-                    icon: "error",
-                    iconColor: "#DB524E",
-                    title: "처리 중 오류가 발생했습니다 (" + error.response?.status + ")",
-                    text: "문제가 지속될 경우 고객센터로 문의 바랍니다.",
-                    showCancelButton: false,
-                    confirmButtonColor: '#1D7151',
-                    confirmButtonText: '확인'
-                })
-            }
         }
     }
 }
