+++ src/assets/css/mypage.css
@@ -19,3 +19,26 @@
 .sidebar-submenu:focus {
     text-decoration: underline !important;
 }
+
+.verifying {
+    display: flex!important;
+    flex-direction: column;
+    align-items: center!important;
+    justify-content: center!important;
+    text-align: center;
+    width: 100%;
+}
+
+.orgImg {
+    height: 369px;
+}
+
+#orgInfo {
+    border-radius: .5rem;
+    background-color: #ffffff;
+    padding: 1.5rem;
+    box-shadow:     0 5px 10px rgba(0, 0, 0, 0.1),
+    0 20px 20px rgba(0, 0, 0, 0.05);
+    margin-top: 1rem;
+    width: 50%;
+}
