+++ src/pages/org/MyPageOrg.js
@@ -26,7 +26,7 @@ function MyPageOrg() {
     }
 
     const navigateEditOrgInfo = () => {
-        navigate('edit');
+        navigate('confirmPwd');
     }
 
     const navigateRequestWithdraw = () => {
