+++ src/pages/org/OrgWithdraw.js
@@ -3,12 +3,10 @@ function OrgWithdraw() {
         <>
             <div className="admin-main">
                 <div className="admin-title">
-                    <h1 class="text-primary">기부처 탈퇴 요청 페이지</h1>
+                    <h1 class="text-primary">기부처 탈퇴</h1>
                 </div>
                 <div>
-                    일단은 메일로 신청해주세요.
-                    폼은 만들지 말지 고민.
-                    만들면 테이블 더 만들어야함.
+                    <h3>Replanet에서의 모금 활동을 중단하시려면 contact@replanetdonation.org로 문의 바랍니다.</h3>
                 </div>
             </div>
         </>
